Create Circle Delight
I wanted to try to create a very trippy look to my project, I was unable to create the game I wanted, so instead I try to make I very hallucinating program animation,I played around with the bouncing balls, gravity, arrays, and growing technique. I was quite disappointed that I couldn't figure out how to make the games and the score to go along with it, I wanted to at least create the mouse to be able to touch the balls and making it bounce higher.